Clean unused lines
