+++ index.html
@@ -1,33 +1,35 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>books</title>
-    <link rel="stylesheet" href="styles.css" />
-  </head>
-  <body>
-    <header>
-      <nav class="navbar">
-        <a href="#list" id="heading" class="nav-brand gradient-style">All awesome books</a>
-        <div class="links">
-          <ul class="nav-menu">
-            <li class="nav-item">
-              <a id="link" class="nav-link" href="#list">List</a>|
-            </li>
-            <li class="nav-item">
-              <a id="add-link" class="nav-link" href="#add-new">Add New</a>|
-            </li>
-            <li class="nav-item">
-              <a id="contact-link" class="nav-link" href="#contact-list">Contact</a>|
-            </li>
-          </ul>
-        </div>
-      </nav>
-    </header>
 
-    <main>
+<head>
+  <meta charset="UTF-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>books</title>
+  <link rel="stylesheet" href="index.css" />
+</head>
+
+<body>
+  <header>
+    <nav class="navbar">
+      <a href="#list" id="heading" class="nav-brand gradient-style">All awesome books</a>
+      <div class="links">
+        <ul class="nav-menu">
+          <li class="nav-item">
+            <a id="link" class="nav-link" href="#list">List</a>|
+          </li>
+          <li class="nav-item">
+            <a id="add-link" class="nav-link" href="#add-new">Add New</a>|
+          </li>
+          <li class="nav-item">
+            <a id="contact-link" class="nav-link" href="#contact-list">Contact</a>|
+          </li>
+        </ul>
+      </div>
+    </nav>
+  </header>
+
+  <main>
     <div class="date">Date</div>
     <br />
     <section class="container">
@@ -63,9 +65,10 @@
       </ul>
     </div>
   </main>
-    <footer>
-      <h3>Copyright @ Awesome Books store</h3>
-    </footer>
-    <script src="main.js"></script>
-  </body>
+  <footer>
+    <h3>Copyright @ Awesome Books store</h3>
+  </footer>
+  <script type="module" src="index.js"></script>
+</body>
+
 </html>
